FE API: move productsByCatnums endpoint from project-base to frontend-api package
