+++ project-base/app/src/Model/Product/ProductRepository.php
@@ -186,41 +186,4 @@ class ProductRepository extends BaseProductRepository
             ->setParameter('variantTypeMain', Product::VARIANT_TYPE_MAIN)
             ->setParameter('domainId', $domainId);
     }
-
-    /**
-     * @param string[] $catnums
-     * @return int[]
-     */
-    public function getProductIdsByCatnums(array $catnums): array
-    {
-        $result = $this->em->createQueryBuilder()
-            ->select('p.id, p.catnum')
-            ->from(Product::class, 'p')
-            ->where('p.catnum IN (:catnums)')
-            ->setParameter('catnums', $catnums)
-            ->getQuery()
-            ->getScalarResult();
-
-        $catnumToIdMap = array_column($result, 'id', 'catnum');
-
-        return $this->getSortedProductIds($catnums, $catnumToIdMap);
-    }
-
-    /**
-     * @param string[] $catnumsToSortBy
-     * @param array<string, int> $catnumToIdMap
-     * @return int[]
-     */
-    private function getSortedProductIds(array $catnumsToSortBy, array $catnumToIdMap): array
-    {
-        $sortedProductIds = [];
-
-        foreach ($catnumsToSortBy as $catnum) {
-            if (array_key_exists($catnum, $catnumToIdMap)) {
-                $sortedProductIds[] = $catnumToIdMap[$catnum];
-            }
-        }
-
-        return $sortedProductIds;
-    }
 }
